infra/requirements.txt
attrs==23.2.0 ; python_version >= "3.12" and python_version < "3.13"
aws-cdk-asset-awscli-v1==2.2.202 ; python_version >= "3.12" and python_version < "3.13"
aws-cdk-asset-kubectl-v20==2.1.2 ; python_version >= "3.12" and python_version < "3.13"
aws-cdk-asset-node-proxy-agent-v6==2.0.3 ; python_version >= "3.12" and python_version < "3.13"
aws-cdk-lib==2.146.0 ; python_version >= "3.12" and python_version < "3.13"
cattrs==23.2.3 ; python_version >= "3.12" and python_version < "3.13"
constructs==10.3.0 ; python_version >= "3.12" and python_version < "3.13"
importlib-resources==6.4.0 ; python_version >= "3.12" and python_version < "3.13"
jsii==1.100.0 ; python_version >= "3.12" and python_version < "3.13"
publication==0.0.3 ; python_version >= "3.12" and python_version < "3.13"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "3.13"
six==1.16.0 ; python_version >= "3.12" and python_version < "3.13"
typeguard==2.13.3 ; python_version >= "3.12" and python_version < "3.13"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "3.13"
